update ibex-client and use redis for auth cache
