+++ .envrc
@@ -14,6 +14,12 @@ export DB_POOL_MIN=1
 export DB_POOL_MAX=5
 export DB_DEBUG=false
 
+export REDIS_0_DNS=${DOCKER_HOST_IP}
+export REDIS_0_PORT=6379
+export REDIS_MASTER_NAME="mymaster"
+export REDIS_PASSWORD=""
+export REDIS_TYPE="standalone"
+
 # add your HONEYCOMB_API_KEY and HONEYCOMB_DATASET in .envrc.secrets to test HONEYCOMB
 if [ -f .envrc.secrets ]; then
   . ./.envrc.secrets
