+++ realtime/package.json
@@ -37,7 +37,9 @@
     "ccxt": "^4.1.70",
     "dotenv": "^16.0.3",
     "grpc-health-check": "^2.0.0",
-    "ibex-client": "^1.0.1",
+    "ibex-client": "^2.1.0",
+    "ioredis": "^5.5.0",
+    "ioredis-cache": "^2.0.3",
     "js-yaml": "^4.1.0",
     "lodash.mergewith": "^4.6.2",
     "node-cache": "^5.1.2",
